+++ src/app/interfaces/mockup-logos.ts
@@ -1,0 +1,5 @@
+export const ABARTH = 'https://cdn.freebiesupply.com/logos/large/2x/abarth-logo-png-transparent.png';
+export const AUDI = 'https://pluspng.com/img-png/audi-logo-png-1920x1080-hd-1080p-1920.png';
+export const KIA = 'https://i0.wp.com/www.maggioraoffroadarena.it/wp-content/uploads/2019/08/kia-logo.png?fit=756%2C387&ssl=1';
+export const OPEL = 'https://www.cavallomotorigroup.it/wp-content/uploads/2018/09/Opel-logo-2002-2048x2048.png';
+export const SEAT = 'https://www.metromaffesa.com/wp-content/uploads/2018/01/adhesivo-seat-logo-efecto-3d-impreso.jpg.png';